fixed a use of applicationDelay that was being stored in an int instead of a long

+++ web_module/branches/reviewer_signup_changes_2008_07_09/src/main/com/topcoder/apps/review/rboard/RBoardApplicationBean.java
@@ -711,7 +711,7 @@ public class RBoardApplicationBean extends BaseEJB {
 						  + timeStampToString(opensOn));
         }
 
-	int applicationDelay = getApplicationDelay(conn, userId);
+	long applicationDelay = getApplicationDelay(conn, userId);
         if (System.currentTimeMillis() < opensOn.getTime() + applicationDelay) {
             throw new RBoardRegistrationException("Sorry, you can not apply for this review yet.  "
 						  + "You will need to wait until "
